refa: Repackage internal/listener
