+++ flux-node/common/mock.go
@@ -2,7 +2,7 @@ package common
 
 import (
 	"github.com/bytepowered/flux/flux-node"
-	"github.com/bytepowered/flux/flux-node/internal"
+	"github.com/bytepowered/flux/flux-node/listener"
 	"github.com/labstack/echo/v4"
 	"net/http/httptest"
 )
@@ -12,7 +12,7 @@ var mock = echo.New()
 func MockWebContext(id string) flux.ServerWebContext {
 	mr := httptest.NewRequest("GET", "http://mocking/"+id, nil)
 	mw := httptest.NewRecorder()
-	return internal.NewServeWebContext(mock.NewContext(mr, mw), id, nil)
+	return listener.NewServeWebContext(mock.NewContext(mr, mw), id, nil)
 }
 
 func MockContext(id string) *flux.Context {
