+++ flux-node/listener/context.go
@@ -1,4 +1,4 @@
-package internal
+package listener
 
 import (
 	"bytes"
@@ -12,6 +12,14 @@ import (
 )
 
 var _ flux.ServerWebContext = new(AdaptWebContext)
+
+// 用于隐藏内部实现的Key
+type webContextKey string
+
+var (
+	keyRequestId  = webContextKey("inter.context.request.id/926820fa-7ad8-4444-9080-d690ce31c93a")
+	keyWebContext = webContextKey("inter.context.web.ctx/890b1fa9-93ad-4b44-af24-85bcbfe646b4")
+)
 
 func NewServeWebContext(ctx echo.Context, reqid string, listener flux.WebListener) flux.ServerWebContext {
 	return &AdaptWebContext{
