+++ flux-node/listener/listener.go
@@ -1,4 +1,4 @@
-package internal
+package listener
 
 import (
 	"context"
